+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/LedgerSpec.hs
@@ -6,6 +6,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
 
 module Test.Cardano.Ledger.Conway.Imp.LedgerSpec (spec) where
 
@@ -13,31 +14,20 @@ import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Conway (
   hardforkConwayBootstrapPhase,
-  hardforkConwayDisallowUnelectedCommitteeFromVoting,
  )
 import Cardano.Ledger.Conway.Core
-import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Conway.PParams (ConwayEraPParams (..))
 import Cardano.Ledger.Conway.Rules (
-  ConwayGovPredFailure (UnelectedCommitteeVoters),
   ConwayLedgerPredFailure (..),
-  ConwayUtxoPredFailure (BadInputsUTxO),
  )
 import Cardano.Ledger.Credential (Credential (..))
 import Cardano.Ledger.DRep
 import Cardano.Ledger.Plutus (SLanguage (..), hashPlutusScript)
-import Cardano.Ledger.Shelley.API.Mempool (ApplyTxError (..), applyTx, mkMempoolEnv)
 import Cardano.Ledger.Shelley.LedgerState
-import Control.Monad.Reader (asks)
-import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
-import qualified Data.Text as T
 import Data.Word (Word32)
-import GHC.Exts (fromList)
-import Lens.Micro ((&), (.~), (<>~), (^.))
-import Lens.Micro.Mtl (use)
+import Lens.Micro ((&), (.~), (^.))
 import Test.Cardano.Ledger.Conway.ImpTest
-import Test.Cardano.Ledger.Core.Rational (IsRatio (..))
 import Test.Cardano.Ledger.Imp.Common
 import Test.Cardano.Ledger.Plutus.Examples (
   alwaysSucceedsNoDatum,
@@ -209,90 +199,3 @@ spec = do
     submitTx_ $
       mkBasicTx $
         mkBasicTxBody & withdrawalsTxBodyL .~ Withdrawals [(ra, mempty)]
-
-  describe "Mempool" $ do
-    let
-      submitFailingMempoolTx cause tx expectedFailures = do
-        globals <- use impGlobalsL
-        nes <- use impNESL
-        slotNo <- use impCurSlotNoG
-        let
-          mempoolEnv = mkMempoolEnv nes slotNo
-          ls = nes ^. nesEsL . esLStateL
-        txFixed <- (tx &) =<< asks iteFixup
-        logToExpr txFixed
-        case applyTx globals mempoolEnv ls txFixed of
-          Left err -> do
-            err `shouldBe` ApplyTxError @era expectedFailures
-          Right _ ->
-            assertFailure $ "Expected failure due to " <> cause <> ": " <> show txFixed
-        pure txFixed
-      submitFailingMempoolTx_ c t = void . submitFailingMempoolTx c t
-
-    it "Duplicate transactions" $ do
-      let
-        newInput = do
-          addr <- freshKeyAddr_
-          amount <- Coin <$> choose (2_000_000, 8_000_000)
-          sendCoinTo addr amount
-
-      inputsCommon <- replicateM 5 newInput
-      inputs1 <- replicateM 2 newInput
-      inputs2 <- replicateM 3 newInput
-
-      txFinal <-
-        submitTx $
-          mkBasicTx $
-            mkBasicTxBody & inputsTxBodyL <>~ fromList (inputsCommon <> inputs1)
-
-      impAnn "Identical transaction" $ do
-        withNoFixup $
-          submitFailingMempoolTx_ "duplicate transaction" txFinal $
-            pure . injectFailure . ConwayMempoolFailure $
-              "All inputs are spent. Transaction has probably already been included"
-
-      impAnn "Overlapping transaction" $ do
-        let txOverlap = mkBasicTx $ mkBasicTxBody & inputsTxBodyL <>~ fromList (inputsCommon <> inputs2)
-        submitFailingMempoolTx_
-          "overlapping transaction"
-          txOverlap
-          [injectFailure $ BadInputsUTxO $ fromList inputsCommon]
-
-    it "Unelected Committee voting" $ whenPostBootstrap $ do
-      _ <- registerInitialCommittee
-      ccCold <- KeyHashObj <$> freshKeyHash
-      curEpochNo <- getsNES nesELL
-      let action =
-            UpdateCommittee
-              SNothing
-              mempty
-              (Map.singleton ccCold (addEpochInterval curEpochNo (EpochInterval 7)))
-              (1 %! 1)
-      proposal <- mkProposal action
-      submitTx_ $
-        mkBasicTx (mkBasicTxBody & proposalProceduresTxBodyL .~ [proposal])
-      ccHot <- registerCommitteeHotKey ccCold
-      govActionId <- do
-        rewardAccount <- registerRewardAccount
-        submitTreasuryWithdrawals [(rewardAccount, Coin 1)]
-
-      let
-        tx =
-          mkBasicTx $
-            mkBasicTxBody
-              & votingProceduresTxBodyL
-                .~ VotingProcedures
-                  ( Map.singleton
-                      (CommitteeVoter ccHot)
-                      (Map.singleton govActionId (VotingProcedure VoteYes SNothing))
-                  )
-      pv <- getProtVer
-      if hardforkConwayDisallowUnelectedCommitteeFromVoting pv
-        then
-          submitFailingTx tx [injectFailure $ UnelectedCommitteeVoters [ccHot]]
-        else do
-          txFixed <-
-            submitFailingMempoolTx "unallowed votes" tx $
-              pure . injectFailure . ConwayMempoolFailure $
-                "Unelected committee members are not allowed to cast votes: " <> T.pack (show (pure @[] ccHot))
-          withNoFixup $ submitTx_ txFixed